% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEmbScatCyclic.R
\name{cyclic_legend}
\alias{cyclic_legend}
\title{Get the cyclic legend}
\arguments{
\item{hue.colors}{BAAA}

\item{hue.n}{BAAA}

\item{alpha}{AAAA}

\item{y.inner}{BAAA}

\item{y.outer}{BAAA}

\item{y.text}{BAAA}

\item{ymax}{BAAA}

\item{text.size}{BAAA}
}
\value{
A ggplot object
}
\description{
BALABALA
}
\details{
BALA
}
\examples{
cyclic_legend()
}
\author{
Shijie C. Zheng
}
