% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/estimate_cycle_stage.R
\name{estimate_cycle_stage}
\alias{estimate_cycle_stage}
\alias{estimate_cycle_stage,ANY-method}
\alias{estimate_cycle_stage,SummarizedExperiment-method}
\alias{estimate_cycle_stage,SingleCellExperiment-method}
\title{Assign cell cycle stages}
\usage{
estimate_cycle_stage(x, ...)

\S4method{estimate_cycle_stage}{ANY}(
  x,
  cycleGene.l = NULL,
  gname = NULL,
  gname.type = c("ENSEMBL", "SYMBOL"),
  species = c("mouse", "human"),
  AnnotationDb = NULL,
  batch.v = NULL,
  corThres = 0.2,
  tolerance = 0.3
)

\S4method{estimate_cycle_stage}{SummarizedExperiment}(x, ..., exprs_values = "logcounts", batch.v = NULL)

\S4method{estimate_cycle_stage}{SingleCellExperiment}(
  x,
  ...,
  exprs_values = "logcounts",
  batch.v = NULL,
  altexp = NULL
)
}
\arguments{
\item{x}{A numeric matrix of **log-expression** values where rows are features and columns are cells.
Alternatively, a \linkS4class{SummarizedExperiment} or \linkS4class{SingleCellExperiment} containing such a matrix.}

\item{...}{For the \code{estimate_cycle_stage} generic, the following additional arguments to pass.
For the \linkS4class{SummarizedExperiment} and  \linkS4class{SingleCellExperiment} methods, additional arguments to pass to the ANY method.}

\item{cycleGene.l}{A list contains the marker genes for each stage. The stage names should be included as names of the elements. If user feed custom list,
they should make sure that the same gene id type for \code{x} and \code{cycleGene.l}. If not custom list is given, \code{\link{RevelioGeneList}} will be used.
Default: NULL}

\item{gname}{Alternative rownames of \code{x}. If provided, this will be used to map genes within \code{x} with genes in \code{ref.m}.
If not provided, the rownames of \code{x} will be used instead. Default: NULL}

\item{gname.type}{The type of gene names as in \code{gname} or rownames of \code{x}. It can be either 'ENSEMBL' or 'SYMBOL'. If the user uses
custom \code{ref.m}, this value will have no effect. Default: 'ENSEMBL'}

\item{species}{The type of species in \code{x}. It can be either 'mouse' or 'human'. If the user uses
custom \code{ref.m}, this value will have no effect. Default: 'mouse'}

\item{AnnotationDb}{An AnnotationDb objects. It is used to map ENSEMBL IDs to gene SYMBOLs.
If no AnnotationDb object being given, the function will use \code{\link[org.Hs.eg.db]{org.Hs.eg.db}} or \code{\link[org.Mm.eg.db]{org.Mm.eg.db}} for human and mouse respetively.}

\item{batch.v}{A string specifies which column in colData of \linkS4class{SummarizedExperiment} or \linkS4class{SingleCellExperiment} to use as the batch variable.
Or it can be a vector, of which the number of elements equals to the number of columns of \code{x}. The 5 stage cell cycle assignments are preformed for each batch separately.
No \code{NA} is permitted. Default: NULL}

\item{corThres}{For each batch and each stage, correlations between expression of each gene and the mean of all genes belonging to that stage
will be calculated to filter the final gene list used for inference. The genes with a correlation between \code{corThres} will not be used for calculating \emph{z}-scores.
Default: 0.2}

\item{tolerance}{For each cell, the function will compare the largest two \emph{z}-scores. If the difference between those two \emph{z}-scores is less than \code{tolerance},
the cell will be treated un-assignable with \code{NA} value returned for that cell. Default: 0.3}

\item{exprs_values}{Integer scalar or string indicating which assay of \code{x} contains the **log-expression** values, which will be used for projection.
If the projection already exists, you can ignore this value. Default: 'logcounts'}

\item{altexp}{String or integer scalar specifying an alternative experiment containing the **log-expression** data, which will be used for projection.
If the projection is already calculated and stored in the \linkS4class{SingleCellExperiment} as a dimred, leave this value to default NULL.}
}
\value{
If the input is a numeric matrix, the discretized cell cycle stages - a factor vector corresponding to each cell will be returned.

If the input is \linkS4class{SummarizedExperiment}, the original \linkS4class{SummarizedExperiment} with the discretized cell cycle stages stored in colData with name 'CCStage' will be returned.

If the input is \linkS4class{SingleCellExperiment}, the original \linkS4class{SingleCellExperiment} with the discretized cell cycle stages stored in colData with name 'CCStage' will be returned.
}
\description{
Assign cell cycle stages using \code{\link{RevelioGeneList}}(Whitfield 2002 list) or user given gene list.
}
\details{
The function will assign the cell to a discretized cell cycle stage by comparing the \emph{z}-scores calculated for each stage markers.
This is modified stage assignment method modified from the method in Schwabe et al.(2020).
Without cycleGene.l input, \code{\link{RevelioGeneList}} will be used.
}
\examples{
neurosphere_example <- estimate_cycle_stage(neurosphere_example, gname.type = "ENSEMBL", species = "mouse")
neurosphere_example2 <- estimate_cycle_stage(neurosphere_example, batch.v = "sample")
neurosphere_example3 <- estimate_cycle_stage(neurosphere_example, batch.v = neurosphere_example$sample)
neurosphere_example <- project_cycle_space(neurosphere_example)
plot(reducedDim(neurosphere_example, "tricycleEmbedding"), col = neurosphere_example$CCStage)
}
\references{
Schwabe D, et al.
\emph{The transcriptome dynamics of single cells during the cell cycle.}
Molecular Systems Biology (2020) 16: e9946
doi:\href{https://doi.org/10.15252/msb.20209946}{10.15252/msb.20209946}.

Zheng SC, et al.
\emph{Universal prediction of cell cycle position using transfer learning.}
}
\author{
Shijie C. Zheng
}
