% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_sce-data.R
\docType{data}
\name{example_sce}
\alias{example_sce}
\title{Example SingleCellExperiment dataset}
\format{
A \linkS4class{SingleCellExperiment} object of 1500 genes and 400 cells.
}
\usage{
data(example_sce)
}
\description{
This a subset of mouse Neurosphere data. 200 cells from sample AX1 and AX2 were randonly 
sampled from the full data. All genes in the GO cell cycle gene list and \code{\link{RevelioGeneList}}
 as well as other random 573 genes were included.
}
\examples{
data(example_sce)
}
\keyword{datasets}
