% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLoessTheta.R
\name{fitLoessTheta}
\alias{fitLoessTheta}
\title{Fit loess line with cell cycle time as predictor}
\arguments{
\item{theta.v}{The cell cycle time - a numeric vector with range between 0 to 2pi.}

\item{y}{The response variable - a numeric vector with the same length as \code{theta.v}.}

\item{span}{The parameter \eqn{\alpha} which controls the degree of smoothing. See \code{\link[stats]{loess}}. Default: 0.3}

\item{length.out}{The number of data points on the fitted lines to be output in the prediction data.frame. Default: 200}

\item{...}{Other arguments input to \code{\link[stats]{loess}}.}
}
\value{
A list with the following elements:
\itemize{
  \item fitted - The fitted vaues on the loess line. A vector of the length of y.
  \item residual - The residual values from the fitted loess line, i.e. y - y.fit. A vector of the length of y.
  \item pred.df - The prediction \code{data.frame} by uniformly sampling theta from 0 - 2*pi. Names of variables: \code{x} and \code{y}. The number of rows equals to \code{length.out}.
  \item loess.o The fitted loess object.
}
}
\description{
The function will fit a loess line using cell cycle time and other variables, such as expression levels of a gene
 or log-transformed totalUMIs numbers. The circular nature of cell cycle time is taken into account by making 3 copies inside the function.
}
\details{
This function fit a normal loess line, but take the circularity of cell cycle time into account by making \code{theta.v} 3 periods
(\code{c(theta.v - 2 * pi, theta.v, theta.v + 2 * pi)}) and repeating y 3 times. Only the fitted values corresponding to original \code{theta.v} 
will be returned. User can use \code{pred.df} to visualize the loess line.
}
\examples{
example_sce <- inferCCTime(example_sce)
top2a.idx <- which(rowData(example_sce)$Gene == 'Top2a')
fit.l <- fitLoessTheta(example_sce$CCTime, assay(example_sce, 'logcounts')[top2a.idx, ])
plot(example_sce$CCTime, assay(example_sce, 'logcounts')[top2a.idx, ])
lines(fit.l$pred.df)
}
\seealso{
\code{\link{inferCCTime}}, for inferring cell cycle time.
}
\author{
Shijie C. Zheng
}
