% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroRef-data.R
\docType{data}
\name{neuroRef}
\alias{neuroRef}
\title{Pre-learned reference projection matrix from Neurosphere dataset}
\format{
An object of class `"data.frame"`
}
\usage{
data(neuroRef)
}
\description{
Data from BALABALA
}
\examples{
data(neuroRef)
}
\keyword{datasets}
