% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaGoCC.R
\name{pcaGoCC}
\alias{pcaGoCC}
\alias{pcaGoCC,SingleCellExperiment-method}
\title{Run PCA on Gene Ontology cell cycle genes}
\usage{
\S4method{pcaGoCC}{SingleCellExperiment}(
  sce.o,
  exprs_values = "logcounts",
  gname.type = c("ENSEMBL", "SYMBOL"),
  species = c("mouse", "human"),
  ntop = 500,
  ncomponents = 20,
  runSeuratBy = NULL,
  nfeatures = 500,
  anchor.features = 2000,
  name = "PCA"
)
}
\arguments{
\item{sce.o}{A \linkS4class{SingleCellExperiment} contains library size normalized **log-expression** matrix.}

\item{exprs_values}{Integer scalar or string indicating which assay of \code{sce.o} contains the **log-expression** values, which will be used to run PCA.
Default: 'logcounts'}

\item{gname.type}{The type of gene names as in \code{gname} or rownames of \code{sce.o}. It can be either 'ENSEMBL' or 'SYMBOL'. Default: 'ENSEMBL'}

\item{species}{The type of species in \code{sce.o}. It can be either 'mouse' or 'human'. The corresponding AnnotationDb object will be
\code{\link[org.Mm.eg.db]{org.Mm.eg.db}} and \code{\link[org.Hs.eg.db]{org.Hs.eg.db}}. Default: 'mouse'}

\item{ntop}{The number of genes with highest variance to use when calculating PCA, as in \code{\link[scater]{calculatePCA}}. Default: 500}

\item{ncomponents}{The number of component components to obtain, as in \code{\link[scater]{calculatePCA}}. Default: 20}

\item{runSeuratBy}{If the value is not NULL, the subsetted data (only GO cell cycle genes) will be integrated by Seurat. See \code{\link[Seurat]{IntegrateData}}.
This value should be set as a column name in \code{colData(sce.o)}. Such as 'sample' or 'batch'. Default: NULL}

\item{nfeatures}{The number of highly variable features within each sample/batch, which will only be used when integrating with Seurat.
See \code{\link[Seurat]{FindVariableFeatures}}. Default: 500}

\item{anchor.features}{The number of genes used in anchor finding during Seurat integration. See \code{\link[Seurat]{FindIntegrationAnchors}}. Default: 2000}

\item{name}{String specifying the name to be used to store the result in the \code{\link[SingleCellExperiment]{reducedDims}} of the output. Default: 'PCA'}

\item{gname}{Alternative rownames of \code{sce.o}. If provided, this will be used to map genes within Gene Ontology cell cycle gene list.
If not provided, the rownames of \code{sce.o} will be used instead. Default: NULL}
}
\value{
A subset \linkS4class{SingleCellExperiment} object with only GO cell cycle genes will be return.
The PCA resulting will be save in reducedDims with chosen name \code{\link[SingleCellExperiment]{reducedDims}(..., name)}.
If Seurat integration is performed, another reducedDims with name 'matched.'+\code{name} will also be included in the \linkS4class{SingleCellExperiment}.
}
\description{
Run PCA on Gene Ontology cell cycle genes abd get a new \linkS4class{SingleCellExperiment}.
User could use this function to learn new reference projection matrix.
}
\details{
The function require an output of a \linkS4class{SingleCellExperiment} object which contains the library size normalized **log-expression** matrix. The full dataset will
be subsetted to genes in the Gene Ontology cell cycle gene list (GO:0007049). The corresponding AnnotationDb object will be
 \code{\link[org.Mm.eg.db]{org.Mm.eg.db}} and \code{\link[org.Hs.eg.db]{org.Hs.eg.db}} for mouse and human respectively. If \code{runSeuratBy} is set, the data will be
 integrated to remove batch effect between samples/batches by Seurat.

User can use this function to make new reference projection matrix by getting the 'rotation' attribute in PCA results. Such as
\code{attr(reducedDim(sce.o, 'PCA'), 'rotation')[, 1:2]}. See examples for more details.
}
\examples{
gocc_sce.o <- pcaGoCC(neurosphere_example)
new.ref <- attr(reducedDim(gocc_sce.o, "PCA"), "rotation")[, seq_len(2)]
neurosphere_example <- inferCCTime(neurosphere_example) ### Use internal NeuroRef to project and infer CC time

### Use new reference to project and infer CC time
new_sce <- inferCCTime(neurosphere_example, ref.m = new.ref, dimred = "ccProjection2")
plot(neurosphere_example$CCTime, new_sce$CCTime)
}
\author{
Shijie C. Zheng
}
