% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotEmbScatCyclic.R
\name{plotEmbScatCyclic}
\alias{plotEmbScatCyclic}
\alias{plotEmbScatCyclic,SingleCellExperiment-method}
\title{Plot embedding with cyclic cell cycle time}
\usage{
plotEmbScatCyclic(sce.o, ...)

\S4method{plotEmbScatCyclic}{SingleCellExperiment}(
  sce.o,
  ...,
  color_by = "CCTime",
  facet_by = NULL,
  dimred = 1,
  dim = 1:2
)
}
\arguments{
\item{sce.o}{A \linkS4class{SingleCellExperiment} contains the embbing to be plotted against.}

\item{color_by}{AAA}

\item{facet_by}{AAA}

\item{dimred}{The name or index of reducedDims in  \linkS4class{SingleCellExperiment} (\code{\link[SingleCellExperiment]{reducedDims}}). Default: 1}

\item{dim}{DBBBBB. Default: 1:2}

\item{fig.title}{The title of the figure. Default: NULL}

\item{point.size}{\code{\link[scattermore]{geom_scattermore}}}

\item{point.alpha}{\code{\link[scattermore]{geom_scattermore}}}

\item{x_lab}{sss}

\item{y_lab}{sss}

\item{hue.colors}{sss}

\item{hue.n}{sss}

\item{plot.legend}{sss}
}
\value{
A ggplot object or a list of ggplot objects.
}
\description{
Generate scat plot of embedding with cyclic cell cycle time or other cyclic variables
}
\details{
BALABALA
}
\examples{
example_sce <- inferCCTime(example_sce)
plotEmbScatCyclic(example_sce, point.size = 3.1, point.alpha = 0.8)
}
\author{
Shijie C. Zheng
}
