% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevelioGeneList-data.R
\docType{data}
\name{RevelioGeneList}
\alias{RevelioGeneList}
\title{5 stage cell cycle gene marker list from Revelio}
\format{
An list balabala
}
\usage{
data(RevelioGeneList)
}
\description{
Data from BALABALA
}
\examples{
data(RevelioGeneList)
}
\keyword{datasets}
