% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_sce-data.R
\docType{data}
\name{example_sce}
\alias{example_sce}
\title{Example SingleCellExperiment dataset}
\format{
An \linkS4class{SingleCellExperiment} object
}
\usage{
data(example_sce)
}
\description{
Data from BALABALA
}
\examples{
data(example_sce)
}
\keyword{datasets}
